requirements.txt
requests>=2.28.0
pandas>=2.0.0
numpy>=1.24.0
vastai>=0.2.0
python-dotenv>=1.0.0
